requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
pydantic>=2.9.0
requests>=2.32.3
yfinance>=0.2.43
feedparser>=6.0.11
pytz>=2024.1
jinja2>=3.1.4